[AIRFLOW-5184] Move GCP Natural Language to core (#5792)
